+++ packages/react/src/hooks/useRequestSignature.ts
@@ -37,10 +37,23 @@ export const useRequestSignature = ({message, address, method, network}: Signatu
     : wc_data && wc_data.error;
   const response: SignatureResponse | undefined = wc_data;
 
-  const requestSignature = () => {
-    if (session && !loading) {
+  const requestSignature = (signatureRequestOverride?: SignatureRequest) => {
+    if (signatureRequestOverride && session && !loading) {
+      log_sdk('useRequestSignature requesting with override...', signatureRequestOverride);
+      return request({
+        topic: session?.topic ?? '',
+        chainId: account ? `${account.network}:${account.chainId}` : 'aleo:1',
+        request: {
+          jsonrpc: '2.0',
+          method: 'requestSignature',
+          params: {
+            ...signatureRequestOverride
+          },
+        },
+      });
+    } else if (session && !loading) {
       log_sdk('useRequestSignature requesting...', [message, address]);
-      request();
+      return request();
     }
   };