+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @puzzlehq/sdk-core
 
+## 0.5.2
+
+### Patch Changes
+
+- Add override to some wallet requests
+
 ## 0.5.1
 
 ### Patch Changes